--- app/src/types/team.ts
@@ -1,6 +1,0 @@
-export interface Team {
-    id: string
-    name: string
-    description: string
-    members?: string[]
-}
